+++ src/core/types/index.ts
@@ -1,4 +1,5 @@
 export * from './block';
 export * from './common';
 export * from './form';
+export * from './formHooks';
 export * from './validation';